pyproject.toml
[project]
name = "check-domains"
version = "0.0.1"
description = "Default template for PDM package"
authors = [{ name = "Dan Pozmanter" }]
dependencies = ["pyyaml>=6.0.2", "python-whois>=0.9.5"]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false
[tool.pdm.dev-dependencies]
test = ["pytest>=7.4.0", "pytest-cov>=4.1.0"]
